+++ NEWS.md
@@ -1,6 +1,7 @@
 26.0 (Not released yet)
 =======================
 
+* Add support for `--xunit-output` in `e3-testsuite-report`.
 * Add support for `--failure-exit-code` in `e3-testsuite-report`.
 
 25.0 (2022-08-31)